Rename module files to *.ts refs #735
