+++ components/add-layers/arcgis/add-layers-arcgis.module.ts
@@ -1,6 +1,7 @@
 import '../../utils/utils.module';
 import addLayersArcgisComponent from './add-layers-arcgis.component';
 import addLayersArcgisService from './add-layers-arcgis.service';
+import * as angular from 'angular';
 
 /**
  * @namespace hs.addLayersArcgis
