+++ components/add-layers/vector/add-layers-vector.module.ts
@@ -1,5 +1,6 @@
 import addLayersVectorService from './add-layers-vector.service';
 import addLayersVectorUrlParserService from './add-layers-vector-url-parser.service';
+import * as angular from 'angular';
 
 /**
  * @namespace hs.addLayersVector
