+++ components/add-layers/wfs/add-layers-wfs.module.ts
@@ -2,6 +2,7 @@ import '../../../common/get-capabilities.module';
 import '../../utils/utils.module';
 import addLayersWfsComponent from './add-layers-wfs.component';
 import capabilitiesErrorDirective from '../capabilities-error.directive';
+import * as angular from 'angular';
 
 /**
  * @namespace hs.addLayersWfs
