+++ components/add-layers/wmts/add-layers-wmts.module.ts
@@ -4,6 +4,7 @@ import addLayersWmtsComponent from './add-layers-wmts.component';
 import addLayersWmtsService from './add-layers-wmts.service';
 import capabilitiesErrorDirective from '../capabilities-error.directive';
 import resampleDialogDirective from '../resample-dialog.directive';
+import * as angular from 'angular';
 
 /**
  * @namespace hs.addLayersWmts
