+++ components/draw/vgi-draw.module.ts
@@ -1,9 +1,10 @@
 import '../core/core.module';
-import '../geolocation/geolocation';
+import '../geolocation/geolocation.module';
 import '../map/map.module';
 import '../utils/utils.module';
 import './draw.module';
 import vgiDrawComponent from './vgi-draw.component';
+import * as angular from 'angular';
 
 /**
  * @namespace hs.vgi-draw
