+++ components/geolocation/geolocation.module.ts
@@ -1,5 +1,6 @@
 import geolocationDirective from './geolocation.directive';
 import geolocationService from './geolocation.service';
+import * as angular from 'angular';
 
 /**
  * @namespace hs.geolocation
