+++ components/layout/index.ts
@@ -10,7 +10,7 @@ import mdSwipeAreaDirective from './md-swipe-area.directive';
 import mdToolbarDirective from './md-toolbar.directive';
 import panelCreatorDirective from './panel-creator.directive';
 import '../core/core.module';
-import '../geolocation/geolocation';
+import '../geolocation/geolocation.module';
 import '../layermanager/layermanager.module';
 import '../map/map.module';
 import '../print/print.module';
